+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/KnowledgeBuilderImpl.java
@@ -15,76 +15,32 @@
 
 package org.drools.compiler.builder.impl;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.Reader;
-import java.io.StringReader;
-import java.io.UncheckedIOException;
-import java.lang.reflect.Array;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.security.AccessController;
-import java.security.PrivilegedAction;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.Stack;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ForkJoinPool;
-import java.util.function.Supplier;
-
 import org.drools.compiler.builder.InternalKnowledgeBuilder;
 import org.drools.compiler.builder.impl.errors.MissingImplementationException;
+import org.drools.compiler.builder.impl.processors.FunctionCompiler;
 import org.drools.compiler.builder.impl.processors.OtherDeclarationProcessor;
 import org.drools.compiler.builder.impl.processors.PackageProcessor;
+import org.drools.compiler.builder.impl.processors.RuleCompiler;
+import org.drools.compiler.builder.impl.processors.RuleValidator;
 import org.drools.compiler.compiler.AnnotationDeclarationError;
-import org.drools.drl.parser.BaseKnowledgeBuilderResultImpl;
 import org.drools.compiler.compiler.ConfigurableSeverityResult;
-import org.drools.drl.extensions.DecisionTableFactory;
 import org.drools.compiler.compiler.DescrBuildError;
-import org.drools.compiler.compiler.Dialect;
 import org.drools.compiler.compiler.DialectCompiletimeRegistry;
-import org.drools.drl.parser.DrlParser;
-import org.drools.drl.parser.DroolsError;
 import org.drools.compiler.compiler.DroolsErrorWrapper;
-import org.drools.drl.parser.DroolsParserException;
 import org.drools.compiler.compiler.DroolsWarning;
 import org.drools.compiler.compiler.DroolsWarningWrapper;
 import org.drools.compiler.compiler.DuplicateFunction;
-import org.drools.compiler.compiler.DuplicateRule;
 import org.drools.compiler.compiler.GlobalError;
-import org.drools.drl.extensions.GuidedRuleTemplateFactory;
-import org.drools.drl.extensions.GuidedRuleTemplateProvider;
 import org.drools.compiler.compiler.PackageBuilderErrors;
 import org.drools.compiler.compiler.PackageBuilderResults;
 import org.drools.compiler.compiler.PackageRegistry;
-import org.drools.drl.parser.ParserError;
 import org.drools.compiler.compiler.ProcessBuilder;
 import org.drools.compiler.compiler.ProcessBuilderFactory;
-import org.drools.drl.extensions.ResourceConversionResult;
 import org.drools.compiler.compiler.ResourceTypeDeclarationWarning;
-import org.drools.compiler.compiler.RuleBuildError;
-import org.drools.compiler.compiler.TypeDeclarationError;
 import org.drools.compiler.compiler.xml.XmlPackageReader;
 import org.drools.compiler.kie.builder.impl.BuildContext;
-import org.drools.drl.parser.lang.ExpanderException;
 import org.drools.compiler.lang.descr.CompositePackageDescr;
-import org.drools.drl.parser.lang.dsl.DSLMappingFile;
-import org.drools.drl.parser.lang.dsl.DSLTokenizedMappingFile;
-import org.drools.drl.parser.lang.dsl.DefaultExpander;
 import org.drools.compiler.rule.builder.RuleBuildContext;
-import org.drools.compiler.rule.builder.RuleBuilder;
 import org.drools.compiler.rule.builder.RuleConditionBuilder;
 import org.drools.compiler.rule.builder.dialect.DialectError;
 import org.drools.core.addon.TypeResolver;
@@ -98,13 +54,11 @@ import org.drools.core.impl.RuleBase;
 import org.drools.core.impl.RuleBaseFactory;
 import org.drools.core.io.impl.BaseResource;
 import org.drools.core.io.impl.ClassPathResource;
-import org.drools.core.io.impl.DescrResource;
 import org.drools.core.io.impl.ReaderResource;
 import org.drools.core.io.internal.InternalResource;
 import org.drools.core.reteoo.CoreComponentFactory;
 import org.drools.core.rule.Function;
 import org.drools.core.rule.ImportDeclaration;
-import org.drools.core.rule.JavaDialectRuntimeData;
 import org.drools.core.rule.Pattern;
 import org.drools.core.rule.TypeDeclaration;
 import org.drools.core.rule.WindowDeclaration;
@@ -133,10 +87,22 @@ import org.drools.drl.ast.descr.RuleDescr;
 import org.drools.drl.ast.descr.TypeDeclarationDescr;
 import org.drools.drl.ast.descr.TypeFieldDescr;
 import org.drools.drl.ast.descr.WindowDeclarationDescr;
+import org.drools.drl.extensions.DecisionTableFactory;
+import org.drools.drl.extensions.GuidedRuleTemplateFactory;
+import org.drools.drl.extensions.GuidedRuleTemplateProvider;
+import org.drools.drl.extensions.ResourceConversionResult;
+import org.drools.drl.parser.BaseKnowledgeBuilderResultImpl;
+import org.drools.drl.parser.DrlParser;
+import org.drools.drl.parser.DroolsError;
+import org.drools.drl.parser.DroolsParserException;
+import org.drools.drl.parser.ParserError;
+import org.drools.drl.parser.lang.ExpanderException;
+import org.drools.drl.parser.lang.dsl.DSLMappingFile;
+import org.drools.drl.parser.lang.dsl.DSLTokenizedMappingFile;
+import org.drools.drl.parser.lang.dsl.DefaultExpander;
 import org.drools.kiesession.rulebase.InternalKnowledgeBase;
 import org.drools.kiesession.rulebase.KnowledgeBaseFactory;
 import org.drools.wiring.api.ComponentsFactory;
-import org.drools.wiring.api.classloader.ProjectClassLoader;
 import org.kie.api.KieBase;
 import org.kie.api.KieBaseConfiguration;
 import org.kie.api.builder.ReleaseId;
@@ -163,8 +129,32 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xml.sax.SAXException;
 
-import static org.drools.compiler.rule.builder.RuleBuildContext.descrToRule;
-import static org.drools.core.impl.KnowledgeBaseImpl.registerFunctionClassAndInnerClasses;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.Reader;
+import java.io.StringReader;
+import java.io.UncheckedIOException;
+import java.lang.reflect.Array;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.Stack;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ForkJoinPool;
+import java.util.function.Supplier;
+
 import static org.drools.core.util.StringUtils.isEmpty;
 import static org.drools.core.util.StringUtils.ucFirst;
 
@@ -969,151 +959,22 @@ public class KnowledgeBuilderImpl implements InternalKnowledgeBuilder {
     }
 
     private void compileFunctions(PackageDescr packageDescr, PackageRegistry pkgRegistry) {
-        List<FunctionDescr> functions = packageDescr.getFunctions();
-        if (!functions.isEmpty()) {
-
-            for (FunctionDescr functionDescr : functions) {
-                if (isEmpty(functionDescr.getNamespace())) {
-                    // make sure namespace is set on components
-                    functionDescr.setNamespace(packageDescr.getNamespace());
-                }
-
-                // make sure functions are compiled using java dialect
-                functionDescr.setDialect("java");
-
-                preCompileAddFunction(functionDescr, pkgRegistry);
-            }
-
-            // iterate and compile
-            for (FunctionDescr functionDescr : functions) {
-                if (filterAccepts(ResourceChange.Type.FUNCTION, functionDescr.getNamespace(), functionDescr.getName())) {
-                    // inherit the dialect from the package
-                    addFunction(functionDescr, pkgRegistry);
-                }
-            }
-
-            // We need to compile all the functions now, so scripting
-            // languages like mvel can find them
-            compileAll();
-
-            for (FunctionDescr functionDescr : functions) {
-                if (filterAccepts(ResourceChange.Type.FUNCTION, functionDescr.getNamespace(), functionDescr.getName())) {
-                    postCompileAddFunction(functionDescr, pkgRegistry);
-                }
-            }
-        }
+        FunctionCompiler functionCompiler = new FunctionCompiler(packageDescr, pkgRegistry, this::filterAccepts);
+        functionCompiler.process();
+        this.results.addAll(functionCompiler.getResults());
     }
 
     private void compileRules(PackageDescr packageDescr, PackageRegistry pkgRegistry) {
-        preProcessRules(packageDescr, pkgRegistry);
-
-        // ensure that rules are ordered by dependency, so that dependent rules are built later
-        SortedRules sortedRules = sortRulesByDependency(packageDescr, pkgRegistry);
-
-        if (!sortedRules.queries.isEmpty()) {
-            compileAllQueries(packageDescr, pkgRegistry, sortedRules.queries);
-        }
-        for (List<RuleDescr> rulesLevel : sortedRules.rules) {
-            compileRulesLevel(packageDescr, pkgRegistry, rulesLevel);
-        }
-    }
-
-    private void compileAllQueries(PackageDescr packageDescr, PackageRegistry pkgRegistry, List<RuleDescr> rules) {
-        Map<String, RuleBuildContext> ruleCxts = buildRuleBuilderContexts(rules, pkgRegistry);
-        for (RuleDescr ruleDescr : rules) {
-            if (filterAccepts(ResourceChange.Type.RULE, ruleDescr.getNamespace(), ruleDescr.getName())) {
-                initRuleDescr(packageDescr, pkgRegistry, ruleDescr);
-                this.results.addAll(addRule(ruleCxts.get(ruleDescr.getName())));
-            }
-        }
+        RuleCompiler ruleCompiler = new RuleCompiler(pkgRegistry, packageDescr, kBase, parallelRulesBuildThreshold,
+                this::filterAccepts, this::filterAcceptsRemoval, packageAttributes, resource, this);
+        ruleCompiler.process();
+        this.results.addAll(ruleCompiler.getResults());
     }
 
     public static class ForkJoinPoolHolder {
         public static final ForkJoinPool COMPILER_POOL = new ForkJoinPool(); // avoid common pool
     }
 
-    private void compileRulesLevel(PackageDescr packageDescr, PackageRegistry pkgRegistry, List<RuleDescr> rules) {
-        boolean parallelRulesBuild = this.kBase == null && parallelRulesBuildThreshold != -1 && rules.size() > parallelRulesBuildThreshold;
-        if (parallelRulesBuild) {
-            Map<String, RuleBuildContext> ruleCxts = new ConcurrentHashMap<>();
-            try {
-                ForkJoinPoolHolder.COMPILER_POOL.submit(() ->
-                                                                rules.stream().parallel()
-                                                                        .filter(ruleDescr -> filterAccepts(ResourceChange.Type.RULE, ruleDescr.getNamespace(), ruleDescr.getName()))
-                                                                        .forEach(ruleDescr -> {
-                                                                            initRuleDescr(packageDescr, pkgRegistry, ruleDescr);
-                                                                            RuleBuildContext context = buildRuleBuilderContext(pkgRegistry, ruleDescr);
-                                                                            ruleCxts.put(ruleDescr.getName(), context);
-                                                                            List<? extends KnowledgeBuilderResult> results = addRule(context);
-                                                                            if (!results.isEmpty()) {
-                                                                                synchronized (this.results) {
-                                                                                    this.results.addAll(results);
-                                                                                }
-                                                                            }
-                                                                        })
-                ).get();
-            } catch (InterruptedException | ExecutionException e) {
-                throw new RuntimeException("Rules compilation failed or interrupted", e);
-            }
-            for (RuleDescr ruleDescr : rules) {
-                RuleBuildContext context = ruleCxts.get(ruleDescr.getName());
-                if (context != null) {
-                    pkgRegistry.getPackage().addRule(context.getRule());
-                }
-            }
-        } else {
-            for (RuleDescr ruleDescr : rules) {
-                if (filterAccepts(ResourceChange.Type.RULE, ruleDescr.getNamespace(), ruleDescr.getName())) {
-                    initRuleDescr(packageDescr, pkgRegistry, ruleDescr);
-                    RuleBuildContext context = buildRuleBuilderContext(pkgRegistry, ruleDescr);
-                    this.results.addAll(addRule(context));
-                    pkgRegistry.getPackage().addRule(context.getRule());
-                }
-            }
-        }
-    }
-
-    private void initRuleDescr(PackageDescr packageDescr, PackageRegistry pkgRegistry, RuleDescr ruleDescr) {
-        if (isEmpty(ruleDescr.getNamespace())) {
-            // make sure namespace is set on components
-            ruleDescr.setNamespace(packageDescr.getNamespace());
-        }
-
-        inheritPackageAttributes(packageAttributes.get(packageDescr.getNamespace()), ruleDescr);
-
-        if (isEmpty(ruleDescr.getDialect())) {
-            ruleDescr.addAttribute(new AttributeDescr("dialect", pkgRegistry.getDialect()));
-        }
-    }
-
-    private List<? extends KnowledgeBuilderResult> addRule(RuleBuildContext context) {
-        return System.getSecurityManager() == null ?
-                internalAddRule(context) :
-                AccessController.<List<? extends KnowledgeBuilderResult>>doPrivileged((PrivilegedAction) () -> internalAddRule(context));
-    }
-
-    private List<? extends KnowledgeBuilderResult> internalAddRule(RuleBuildContext context) {
-        RuleBuilder.build(context);
-
-        context.getRule().setResource(context.getRuleDescr().getResource());
-
-        context.getDialect().addRule(context);
-
-        if (context.needsStreamMode()) {
-            context.getPkg().setNeedStreamMode();
-        }
-
-        if (context.getErrors().isEmpty()) {
-            return context.getWarnings();
-        } else if (context.getWarnings().isEmpty()) {
-            return context.getErrors();
-        }
-
-        List<KnowledgeBuilderResult> result = new ArrayList<>();
-        result.addAll(context.getErrors());
-        result.addAll(context.getWarnings());
-        return result;
-    }
 
     boolean filterAccepts(ResourceChange.Type type, String namespace, String name) {
         return assetFilter == null || !AssetFilter.Action.DO_NOTHING.equals(assetFilter.accept(type, namespace, name));
@@ -1123,243 +984,6 @@ public class KnowledgeBuilderImpl implements InternalKnowledgeBuilder {
         return assetFilter != null && AssetFilter.Action.REMOVE.equals(assetFilter.accept(type, namespace, name));
     }
 
-    private void preProcessRules(PackageDescr packageDescr, PackageRegistry pkgRegistry) {
-        if (this.kBase == null) {
-            return;
-        }
-
-        InternalKnowledgePackage pkg = pkgRegistry.getPackage();
-        boolean needsRemoval = false;
-
-        // first, check if any rules no longer exist
-        for (org.kie.api.definition.rule.Rule rule : pkg.getRules()) {
-            if (filterAcceptsRemoval(ResourceChange.Type.RULE, rule.getPackageName(), rule.getName())) {
-                needsRemoval = true;
-                break;
-            }
-        }
-
-        if (!needsRemoval) {
-            for (RuleDescr ruleDescr : packageDescr.getRules()) {
-                if (filterAccepts(ResourceChange.Type.RULE, ruleDescr.getNamespace(), ruleDescr.getName())) {
-                    if (pkg.getRule(ruleDescr.getName()) != null) {
-                        needsRemoval = true;
-                        break;
-                    }
-                }
-            }
-        }
-
-        if (needsRemoval) {
-            kBase.enqueueModification(() -> {
-                Collection<RuleImpl> rulesToBeRemoved = new HashSet<>();
-
-                for (org.kie.api.definition.rule.Rule rule : pkg.getRules()) {
-                    if (filterAcceptsRemoval(ResourceChange.Type.RULE, rule.getPackageName(), rule.getName())) {
-                        rulesToBeRemoved.add(((RuleImpl) rule));
-                    }
-                }
-
-                rulesToBeRemoved.forEach(pkg::removeRule);
-
-                for (RuleDescr ruleDescr : packageDescr.getRules()) {
-                    if (filterAccepts(ResourceChange.Type.RULE, ruleDescr.getNamespace(), ruleDescr.getName())) {
-                        RuleImpl rule = pkg.getRule(ruleDescr.getName());
-                        if (rule != null) {
-                            rulesToBeRemoved.add(rule);
-                        }
-                    }
-                }
-
-                if (!rulesToBeRemoved.isEmpty()) {
-                    rulesToBeRemoved.addAll(findChildrenRulesToBeRemoved(packageDescr, rulesToBeRemoved));
-                    kBase.removeRules(rulesToBeRemoved);
-                }
-            });
-        }
-    }
-
-    private Collection<RuleImpl> findChildrenRulesToBeRemoved(PackageDescr packageDescr, Collection<RuleImpl> rulesToBeRemoved) {
-        Collection<String> childrenRuleNamesToBeRemoved = new HashSet<>();
-        Collection<RuleImpl> childrenRulesToBeRemoved = new HashSet<>();
-        for (RuleImpl rule : rulesToBeRemoved) {
-            if (rule.hasChildren()) {
-                for (RuleImpl child : rule.getChildren()) {
-                    if (!rulesToBeRemoved.contains(child)) {
-                        // if a rule has a child rule not marked to be removed ...
-                        childrenRulesToBeRemoved.add(child);
-                        childrenRuleNamesToBeRemoved.add(child.getName());
-                        // ... remove the child rule but also add it back to the PackageDescr in order to readd it when also the parent rule will be readded ...
-                        RuleDescr toBeReadded = new RuleDescr(child.getName());
-                        toBeReadded.setNamespace(packageDescr.getNamespace());
-                        packageDescr.addRule(toBeReadded);
-                    }
-                }
-            }
-        }
-        // ... add a filter to the PackageDescr to also consider the readded children rules as updated together with the parent one
-        if (!childrenRuleNamesToBeRemoved.isEmpty()) {
-            ((CompositePackageDescr) packageDescr).addFilter((type, pkgName, assetName) -> childrenRuleNamesToBeRemoved.contains(assetName) ? AssetFilter.Action.UPDATE : AssetFilter.Action.DO_NOTHING);
-        }
-        return childrenRulesToBeRemoved;
-    }
-
-    private Map<String, RuleBuildContext> buildRuleBuilderContexts(List<RuleDescr> rules, PackageRegistry pkgRegistry) {
-        Map<String, RuleBuildContext> map = new HashMap<>();
-        for (RuleDescr ruleDescr : rules) {
-            RuleBuildContext context = buildRuleBuilderContext(pkgRegistry, ruleDescr);
-            map.put(ruleDescr.getName(), context);
-            pkgRegistry.getPackage().addRule(context.getRule());
-        }
-        return map;
-    }
-
-    private RuleBuildContext buildRuleBuilderContext(PackageRegistry pkgRegistry, RuleDescr ruleDescr) {
-        if (ruleDescr.getResource() == null) {
-            ruleDescr.setResource(resource);
-        }
-
-        DialectCompiletimeRegistry ctr = pkgRegistry.getDialectCompiletimeRegistry();
-        RuleBuildContext context = new RuleBuildContext(this,
-                                                        ruleDescr,
-                                                        ctr,
-                                                        pkgRegistry.getPackage(),
-                                                        ctr.getDialect(pkgRegistry.getDialect()));
-        RuleBuilder.preProcess(context);
-        return context;
-    }
-
-    private SortedRules sortRulesByDependency(PackageDescr packageDescr, PackageRegistry pkgRegistry) {
-        // Using a topological sorting algorithm
-        // see http://en.wikipedia.org/wiki/Topological_sorting
-
-        InternalKnowledgePackage pkg = pkgRegistry.getPackage();
-
-        List<RuleDescr> roots = new ArrayList<>();
-        Map<String, List<RuleDescr>> children = new HashMap<>();
-        LinkedHashMap<String, RuleDescr> sorted = new LinkedHashMap<>();
-        List<RuleDescr> queries = new ArrayList<>();
-        Set<String> compiledRules = new HashSet<>();
-
-        for (RuleDescr ruleDescr : packageDescr.getRules()) {
-            if (ruleDescr.isQuery()) {
-                queries.add(ruleDescr);
-            } else if (!ruleDescr.hasParent()) {
-                roots.add(ruleDescr);
-            } else {
-                if (pkg.getRule(ruleDescr.getParentName()) != null) {
-                    // The parent of this rule has been already compiled
-                    compiledRules.add(ruleDescr.getParentName());
-                }
-                children.computeIfAbsent(ruleDescr.getParentName(), k -> new ArrayList<>()).add(ruleDescr);
-            }
-        }
-
-        SortedRules sortedRules = new SortedRules();
-        sortedRules.queries = queries;
-
-        if (children.isEmpty()) { // Sorting not necessary
-            if (!queries.isEmpty()) { // Build all queries first
-                packageDescr.getRules().removeAll(queries);
-                packageDescr.getRules().addAll(0, queries);
-                sortedRules.rules.add(packageDescr.getRules().subList(queries.size(), packageDescr.getRules().size()));
-            } else {
-                sortedRules.rules.add(packageDescr.getRules());
-            }
-            return sortedRules;
-        }
-
-        for (String compiledRule : compiledRules) {
-            List<RuleDescr> childz = children.remove(compiledRule);
-            roots.addAll(childz);
-        }
-
-        List<RuleDescr> rulesLevel = roots;
-        while (!rulesLevel.isEmpty()) {
-            rulesLevel = sortRulesLevel(rulesLevel, sorted, sortedRules, children);
-            sortedRules.newLevel();
-        }
-
-        reportHierarchyErrors(children, sorted);
-
-        packageDescr.getRules().clear();
-        packageDescr.getRules().addAll(queries);
-        for (RuleDescr descr : sorted.values()) {
-            packageDescr.getRules().add(descr);
-        }
-        return sortedRules;
-    }
-
-    private List<RuleDescr> sortRulesLevel(final List<RuleDescr> rulesLevel,
-                                           final LinkedHashMap<String, RuleDescr> sorted, final SortedRules sortedRules,
-                                           final Map<String, List<RuleDescr>> children) {
-        final List<RuleDescr> nextLevel = new ArrayList<>();
-        rulesLevel.forEach(ruleDescr -> {
-            sortedRules.addRule(ruleDescr);
-            sorted.put(ruleDescr.getName(), ruleDescr);
-            final List<RuleDescr> childz = children.remove(ruleDescr.getName());
-            if (childz != null) {
-                nextLevel.addAll(childz);
-            }
-        });
-        return nextLevel;
-    }
-
-    private static class SortedRules {
-
-        List<RuleDescr> queries;
-        final List<List<RuleDescr>> rules = new ArrayList<>();
-        List<RuleDescr> current = new ArrayList<>();
-
-        SortedRules() {
-            newLevel();
-        }
-
-        void addRule(RuleDescr rule) {
-            current.add(rule);
-        }
-
-        void newLevel() {
-            current = new ArrayList<>();
-            rules.add(current);
-        }
-    }
-
-    private void reportHierarchyErrors(Map<String, List<RuleDescr>> parents,
-                                       Map<String, RuleDescr> sorted) {
-        boolean circularDep = false;
-        for (List<RuleDescr> rds : parents.values()) {
-            for (RuleDescr ruleDescr : rds) {
-                if (parents.get(ruleDescr.getParentName()) != null
-                        && (sorted.containsKey(ruleDescr.getName()) || parents.containsKey(ruleDescr.getName()))) {
-                    circularDep = true;
-                    results.add(new RuleBuildError(descrToRule(ruleDescr), ruleDescr, null,
-                                                   "Circular dependency in rules hierarchy"));
-                    break;
-                }
-                manageUnresolvedExtension(ruleDescr, sorted.values());
-            }
-            if (circularDep) {
-                break;
-            }
-        }
-    }
-
-    private void manageUnresolvedExtension(RuleDescr ruleDescr,
-                                           Collection<RuleDescr> candidates) {
-        List<String> candidateRules = new ArrayList<>();
-        for (RuleDescr r : candidates) {
-            if (StringUtils.stringSimilarity(ruleDescr.getParentName(), r.getName(), StringUtils.SIMILARITY_STRATS.DICE) >= 0.75) {
-                candidateRules.add(r.getName());
-            }
-        }
-        String msg = "Unresolved parent name " + ruleDescr.getParentName();
-        if (!candidateRules.isEmpty()) {
-            msg += " >> did you mean any of :" + candidateRules;
-        }
-        results.add(new RuleBuildError(descrToRule(ruleDescr), ruleDescr, msg,
-                                       "Unable to resolve parent rule, please check that both rules are in the same package"));
-    }
 
     private String getPackageDialect(PackageDescr packageDescr) {
         String dialectName = this.defaultDialect;
@@ -1508,57 +1132,8 @@ public class KnowledgeBuilderImpl implements InternalKnowledgeBuilder {
     }
 
     protected void validateUniqueRuleNames(final PackageDescr packageDescr) {
-        final Set<String> names = new HashSet<>();
         PackageRegistry packageRegistry = this.pkgRegistryMap.get(packageDescr.getNamespace());
-        InternalKnowledgePackage pkg = null;
-        if (packageRegistry != null) {
-            pkg = packageRegistry.getPackage();
-        }
-        for (final RuleDescr rule : packageDescr.getRules()) {
-            validateRule(packageDescr, rule);
-
-            final String name = rule.getName();
-            if (names.contains(name)) {
-                addBuilderResult(new ParserError(rule.getResource(),
-                                                 "Duplicate rule name: " + name,
-                                                 rule.getLine(),
-                                                 rule.getColumn(),
-                                                 packageDescr.getNamespace()));
-            }
-            if (pkg != null) {
-                RuleImpl duplicatedRule = pkg.getRule(name);
-                if (duplicatedRule != null) {
-                    Resource resource = rule.getResource();
-                    Resource duplicatedResource = duplicatedRule.getResource();
-                    if (resource == null || duplicatedResource == null || duplicatedResource.getSourcePath() == null ||
-                            duplicatedResource.getSourcePath().equals(resource.getSourcePath())) {
-                        addBuilderResult(new DuplicateRule(rule,
-                                                           packageDescr,
-                                                           this.configuration));
-                    } else {
-                        addBuilderResult(new ParserError(rule.getResource(),
-                                                         "Duplicate rule name: " + name,
-                                                         rule.getLine(),
-                                                         rule.getColumn(),
-                                                         packageDescr.getNamespace()));
-                    }
-                }
-            }
-            names.add(name);
-        }
-    }
-
-    private void validateRule(PackageDescr packageDescr,
-                              RuleDescr rule) {
-        if (rule.hasErrors()) {
-            for (String error : rule.getErrors()) {
-                addBuilderResult(new ParserError(rule.getResource(),
-                                                 error + " in rule " + rule.getName(),
-                                                 rule.getLine(),
-                                                 rule.getColumn(),
-                                                 packageDescr.getNamespace()));
-            }
-        }
+        new RuleValidator(packageRegistry, packageDescr, configuration).process();
     }
 
     void mergePackage(PackageRegistry pkgRegistry, PackageDescr packageDescr) {
@@ -1739,34 +1314,6 @@ public class KnowledgeBuilderImpl implements InternalKnowledgeBuilder {
                 }
             } else {
                 pkgRegistry.getPackage().addWindowDeclaration(window);
-            }
-        }
-    }
-
-    private void addFunction(final FunctionDescr functionDescr, PackageRegistry pkgRegistry) {
-        Dialect dialect = pkgRegistry.getDialectCompiletimeRegistry().getDialect(functionDescr.getDialect());
-        dialect.addFunction(functionDescr,
-                            pkgRegistry.getTypeResolver(),
-                            this.resource);
-    }
-
-    private void preCompileAddFunction(final FunctionDescr functionDescr, PackageRegistry pkgRegistry) {
-        Dialect dialect = pkgRegistry.getDialectCompiletimeRegistry().getDialect(functionDescr.getDialect());
-        dialect.preCompileAddFunction(functionDescr,
-                                      pkgRegistry.getTypeResolver());
-    }
-
-    private void postCompileAddFunction(final FunctionDescr functionDescr, PackageRegistry pkgRegistry) {
-        Dialect dialect = pkgRegistry.getDialectCompiletimeRegistry().getDialect(functionDescr.getDialect());
-        dialect.postCompileAddFunction(functionDescr, pkgRegistry.getTypeResolver());
-
-        if (rootClassLoader instanceof ProjectClassLoader ) {
-            String functionClassName = functionDescr.getClassName();
-            JavaDialectRuntimeData runtime = ((JavaDialectRuntimeData) pkgRegistry.getDialectRuntimeRegistry().getDialectData("java"));
-            try {
-                registerFunctionClassAndInnerClasses(functionClassName, runtime, ((ProjectClassLoader) rootClassLoader)::storeClass);
-            } catch (ClassNotFoundException e) {
-                throw new RuntimeException(e);
             }
         }
     }
@@ -1961,16 +1508,6 @@ public class KnowledgeBuilderImpl implements InternalKnowledgeBuilder {
         return this.rootClassLoader;
     }
 
-    //Entity rules inherit package attributes
-    private void inheritPackageAttributes(Map<String, AttributeDescr> pkgAttributes,
-                                          RuleDescr ruleDescr) {
-        if (pkgAttributes == null) {
-            return;
-        }
-        for (AttributeDescr attrDescr : pkgAttributes.values()) {
-            ruleDescr.getAttributes().putIfAbsent(attrDescr.getName(), attrDescr);
-        }
-    }
 
     private ChangeSet parseChangeSet(Resource resource) throws IOException, SAXException {
         XmlChangeSetReader reader = new XmlChangeSetReader(this.configuration.getSemanticModules());
